Add hooks folder and componetize card parts
