+++ src/components/cacosted/Notification.jsx
@@ -1,0 +1,9 @@
+import React from 'react'
+
+export const Notification = ({copyClass}) => {
+	return (
+		<div className={`absolute top-5 mx-auto text-center border-amber-500 border-2 rounded-lg transition-opacity duration-500 ${copyClass}`}>
+			<span className="text-white p-5 relative block">Copied to clipboard</span>
+		</div>
+	)
+}
