+++ src/pages/entry/cacosted/index.astro
@@ -1,6 +1,6 @@
 ---
 import Layout from '@layout'
-import {Card} from '@components/cacosted/Card.jsx'
+import {Card} from '@components/cacosted/Card'
 ---
 
 <Layout title="cacosted">
